Add base support for list arguments
This is not a final solution for displaying lists of values, as they
will be displayed on a single line, so when a list contains many
elements it wouldn't look great.

All this "info/list displaying" machinery is done and tested in emacs
interface for GNU Guix, so eventually I need to make a stand-alone
package providing this "buffer interface".